code/requirements.txt
geopandas
graphqlclient
matplotlib
networkx
numpy
pandas
requests
ripe.atlas.cousteau
rtree
selenium
shapely
